backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
numpy
google-generativeai

# Nuevas dependencias para datos reales
requests
pandas
aiohttp
scipy
scikit-learn
feedparser
beautifulsoup4